% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{cleanup_figures}
\alias{cleanup_figures}
\title{cleanup_figures}
\usage{
cleanup_figures()
}
\description{
This is a helper function that removes all files from the inst/figures directory.
As the html reports have embeded resources, this is useful for cleaning up the
figures directory between each sequential report iteration.
}
