% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort_checks.R
\name{effort_na.count.type}
\alias{effort_na.count.type}
\title{Check for missing count_quantity values}
\usage{
effort_na.count.type(data)
}
\arguments{
\item{data}{List of creel data from \code{\link[creelutils:fetch_dwg]{creelutils::fetch_dwg()}}.}
}
\value{
list(result, detail)
}
\description{
Check for missing count_quantity values
}
