% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort_checks.R
\name{effort_na.location}
\alias{effort_na.location}
\title{Check for missing effort location}
\usage{
effort_na.location(data)
}
\arguments{
\item{data}{List of creel data from \code{\link[creelutils:fetch_dwg]{creelutils::fetch_dwg()}}.}
}
\value{
list(result, detail)
}
\description{
Check for missing effort location
}
