% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_dataset}
\alias{example_dataset}
\title{Example creel dataset}
\format{
An object of class \code{list} of length 6.
}
\usage{
example_dataset
}
\description{
This dataset is a copy of the "Skagit winter steelhead 2021" fishery. It was downloaded on January 22, 2025 from the
public data repository data.wa.gov. Hosted there are views from WDFW's internal creel database, which are formatted
queries of primary data components (e.g., effort, interviews, and catch).
}
\details{
@format A list with the following data frames:
\describe{
\item{effort}{placeholder with columns \code{count_sequence}, and etc.}
}
@examples
summary(example_dataset)
head(example_dataset$effort)
}
\keyword{datasets}
