% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_all_checks.R
\name{fishapps_link}
\alias{fishapps_link}
\title{Build FishApps event links from creel_event_id}
\usage{
fishapps_link(creel_event_id)
}
\arguments{
\item{creel_event_id}{Character vector of event IDs.}
}
\value{
Character vector of HTML anchor tags.
}
\description{
Build FishApps event links from creel_event_id
}
\keyword{internal}
