% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_datasets.R
\name{format_assessment_gt}
\alias{format_assessment_gt}
\title{Format multi-fishery assessment as a gt stoplight table}
\usage{
format_assessment_gt(results)
}
\arguments{
\item{results}{Tibble from \code{\link[=assess_datasets]{assess_datasets()}}.}
}
\value{
A \code{gt} table object.as
}
\description{
Format multi-fishery assessment as a gt stoplight table
}
