% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interview_checks.R
\name{interview_end.time.interview.time}
\alias{interview_end.time.interview.time}
\title{Check for fishing end time after interview time}
\usage{
interview_end.time.interview.time(data)
}
\arguments{
\item{data}{List of creel data from \code{\link[creelutils:fetch_dwg]{creelutils::fetch_dwg()}}.}
}
\value{
\code{list(result, detail)} where result is
a one-row summary tibble and \code{detail} is flagged records or \code{NULL} on pass.
}
\description{
Check for fishing end time after interview time
}
