% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interview_checks.R
\name{interview_missing.trip.status}
\alias{interview_missing.trip.status}
\title{interview_missing.trip.status}
\usage{
interview_missing.trip.status(data)
}
\arguments{
\item{data}{creel data set via CreelEstimateR::fetch_dwg(params$fishery_name)}
}
\value{
PASS/FAIL with message and indication whether field is critical to estimation
}
\description{
interview_missing.trip.status
}
