% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interview_checks.R
\name{interview_trailer.count.vehicle.count}
\alias{interview_trailer.count.vehicle.count}
\title{Check for trailer count exceeding vehicle count}
\usage{
interview_trailer.count.vehicle.count(data)
}
\arguments{
\item{data}{List of creel data from \code{\link[creelutils:fetch_dwg]{creelutils::fetch_dwg()}}.}
}
\value{
\code{list(result, detail)} where \code{result} is
a one-row summary tibble and \code{detail} is flagged records or \code{NULL} on pass.
}
\description{
Check for trailer count exceeding vehicle count
}
