% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_all_checks.R
\name{create_results_table}
\alias{create_results_table}
\title{create_results_table}
\usage{
create_results_table(pass, critical, qaqc_check_type, error_count, message)
}
\arguments{
\item{pass}{logical / TRUE if the check passed, FALSE if it failed}

\item{critical}{logical / TRUE if the check is critical, FALSE if it is not}

\item{qaqc_check_type}{character / type of QAQC check}

\item{error_count}{numeric / number of errors found for a given check}

\item{message}{character / message to display in results table}
}
\value{
template results table
}
\description{
This helper function creates a results table for each QAQC check.
}
\keyword{internal}
