% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_qaqc_report.R
\name{generate_qaqc_report}
\alias{generate_qaqc_report}
\title{generate_qaqc_report}
\usage{
generate_qaqc_report(fishery_names)
}
\arguments{
\item{fishery_names}{input vector of freshwater creel fishery_names}
}
\value{
HTML QAQC report for each fishery
}
\description{
generate_qaqc_report
}
