% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\title{Generate QAQC report}
\usage{
generate_report(fishery_names, output_dir = NULL)
}
\arguments{
\item{fishery_names}{A character vector of freshwater creel fishery names.}

\item{output_dir}{A character string specifying the directory to save the reports.}
}
\value{
HTML QAQC report for each fishery
}
\description{
This function generates HTML QAQC reports for specified freshwater creel fisheries
and saves them to a specified directory. By default, the reports are saved to a
directory called "CreelDataQAQC_reports" in the current working directory.
}
\examples{
\dontrun{
# Default usage: saves report to current working directory
generate_report(c("Fishery A", "Fishery B"))

# Custom output directory
generate_report(c("Fishery A", "Fishery B"), output_dir = "path/to/output")
}
}
